feat: 🎸 update connect logic for targets with static addresses
✅ Closes: ICU-10430
