+++ ui/desktop/app/routes/scopes/scope/projects/targets.js
@@ -7,15 +7,12 @@ import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
 import { action } from '@ember/object';
 import { resourceFilter } from 'core/decorators/resource-filter';
-import { loading } from 'ember-loading';
 
 export default class ScopesScopeProjectsTargetsRoute extends Route {
   // =services
 
   @service can;
   @service clusterUrl;
-  @service confirm;
-  @service ipc;
   @service resourceFilterStore;
   @service router;
   @service session;
@@ -72,58 +69,6 @@ export default class ScopesScopeProjectsTargetsRoute extends Route {
   @action
   acknowledge(session) {
     session.acknowledged = true;
-  }
-
-  /**
-   * Establish a session to current target.
-   * @param {TargetModel} model
-   * @param {HostModel} host
-   */
-  @action
-  @loading
-  async connect(model, host) {
-    // TODO: Connect: move this logic into the target model
-    try {
-      // Check for CLI
-      const cliExists = await this.ipc.invoke('cliExists');
-      if (!cliExists) throw new Error('Cannot find Boundary CLI.');
-
-      const options = {
-        target_id: model.id,
-        token: this.session.data.authenticated.token,
-      };
-
-      if (host) options.host_id = host.id;
-
-      // Create target session
-      const connectionDetails = await this.ipc.invoke('connect', options);
-
-      // Associate the connection details with the session
-      const { session_id, address, port, credentials } = connectionDetails;
-      const session = await this.store.findRecord('session', session_id);
-
-      // Flag the session has been open in the desktop client
-      session.started_desktop_client = true;
-      // Update the session record with proxy information from the CLI
-      // In the future, it may make sense to push this off to the API so that
-      // we don't have to manually persist the proxy details.
-      session.proxy_address = address;
-      session.proxy_port = port;
-      if (credentials) {
-        credentials.forEach((cred) => session.addCredential(cred));
-      }
-
-      await this.router.transitionTo(
-        'scopes.scope.projects.sessions.session',
-        session
-      );
-    } catch (e) {
-      this.confirm
-        .confirm(e.message, { isConnectError: true })
-        // Retry
-        .then(() => this.connect(model))
-        .catch(() => null /* no op */);
-    }
   }
 
   /**
