+++ ui/desktop/app/templates/scopes/scope/projects/targets/target.hbs
@@ -20,7 +20,7 @@
       @text={{t 'resources.session.actions.connect'}}
       @icon='entry-point'
       @iconPosition='trailing'
-      {{on 'click' (route-action 'connect' @model)}}
+      {{on 'click' (route-action 'preConnect' @model)}}
     />
   </page.actions>
